Refactoring: creating separate folder for components; replacing styles to component folders; importing styles directly to components; replacing all app related scripts, img and styles into "oop" folder
