+++ index.html
@@ -47,50 +47,9 @@
     <link href="https://fonts.googleapis.com/css2?family=Nunito:ital,wght@0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;0,1000;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900;1,1000&display=swap"
           rel="stylesheet">
 
-    <!--STYLES-->
-    <link href="css/main.css" rel="stylesheet" type="text/css">
-    <link href="css/null.css" rel="stylesheet" type="text/css">
-
-    <!--COMPONENT STYLES-->
-    <link href="css/AddTask.css" rel="stylesheet" type="text/css">
-    <link href="css/App.css" rel="stylesheet" type="text/css">
-    <link href="css/Button.css" rel="stylesheet" type="text/css">
-    <link href="css/CustomCheckInput.css" rel="stylesheet" type="text/css">
-    <link href="css/Header.css" rel="stylesheet" type="text/css">
-    <link href="css/Input.css" rel="stylesheet" type="text/css">
-    <link href="css/Modal.css" rel="stylesheet" type="text/css">
-    <link href="css/TaskItem.css" rel="stylesheet" type="text/css">
-    <link href="css/TaskList.css" rel="stylesheet" type="text/css">
-    <link href="css/TaskTag.css" rel="stylesheet" type="text/css">
-    <link href="css/WeatherWidget.css" rel="stylesheet" type="text/css">
-    <link href="css/TodayTasks.css" rel="stylesheet" type="text/css">
-
-
     <title>Quantori Homework. ToDo List</title>
 </head>
 <body>
+<noscript>You need to enable JavaScript to run this app.</noscript>
 </body>
-
-<script src="api/api.js" type="module"></script>
-<script src="api/itemsAPI.js" type="module"></script>
-<script src="api/weatherAPI.js" type="module"></script>
-
-<script src="oop/base_classes.js" type="module"></script>
-<script src="oop/components/AddTask.js" type="module"></script>
-<script src="oop/components/TodayTasks.js" type="module"></script>
-<script src="oop/components/Button.js" type="module"></script>
-<script src="oop/components/CustomCheckInput.js" type="module"></script>
-<script src="oop/components/WeatherWidget.js" type="module"></script>
-<script src="oop/components/Header.js" type="module"></script>
-<script src="oop/components/Input.js" type="module"></script>
-<script src="oop/components/Modal.js" type="module"></script>
-<script src="oop/components/TaskItem.js" type="module"></script>
-<script src="oop/components/TaskList.js" type="module"></script>
-<script src="oop/components/TaskTag.js" type="module"></script>
-
-<script src="oop/components/App.js" type="module"></script>
-
-<script src="oop/index.js" type="module"></script>
-
-
 </html>
