+++ oop/components/AddTask/AddTask.js
@@ -1,8 +1,9 @@
-import {Component} from "../base_classes.js";
-import Input from "./Input.js";
-import CustomCheckInput from "./CustomCheckInput.js";
-import TaskTag from "./TaskTag.js";
-import Button from "./Button.js";
+import {Component} from "../../base_classes.js";
+import Input from "../Input/Input.js";
+import CustomCheckInput from "../CustomCheckInput/CustomCheckInput.js";
+import TaskTag from "../TaskTag/TaskTag.js";
+import Button from "../Button/Button.js";
+import "./AddTask.css"
 
 class AddTask extends Component {
     constructor() {
