+++ oop/components/Modal/Modal.js
@@ -1,4 +1,5 @@
-import {Component} from "../base_classes.js";
+import {Component} from "../../base_classes.js";
+import "./Modal.css"
 
 class Modal extends Component {
     constructor() {
