+++ oop/index.js
@@ -1,3 +1,5 @@
-import App from "./components/App.js";
+import App from "./components/App/App.js";
+import "./css/main.css"
+import "./css/null.css"
 
 document.body.appendChild(new App().render());
